pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "ultraprompt"
version = "0.1.0"
description = "Prompt-guided instance segmentation (SAM2→YOLO) with a clean GUI."
readme = "README.md"
requires-python = "==3.11.*"
authors = [{name="Sumner B. Harris, Anamika Sen"}]
license = {text = "MIT"}
keywords = ["segmentation", "SAM2", "YOLO", "annotation", "Ultralytics"]

dependencies = [
  "ultralytics>=8.3.0",
  "torch>=2.3.1",               # CPU by default; use CUDA index for GPU
  "torchvision>=0.18.1",        # keep paired with your torch version
  "PySide6>=6.6",
  "opencv-python>=4.8.0",
  "Pillow>=10.0.0",
  "numpy>=1.25.0",
  "shapely>=2.0.0",
  "matplotlib>=3.8.0",
]

[project.optional-dependencies]
# Your optional Jupyter tools
notebook = [
  "notebook>=7.0.0",
  "ipykernel>=6.25.0",
  # "jupyterlab>=4.0.0",  # include if you prefer Lab
]
dev = ["pytest", "pytest-cov", "ruff", "mypy", "ipython"]

[project.scripts]
ultraprompt-gui = "ultraprompt.gui:main"

[tool.hatch.build.targets.wheel]
packages = ["src/ultraprompt"]